fix case in path and use absolute paths
Fix the case `rsKeys` vs `rskeys` in the file lookup path
Use absolute paths to compare paths so we can use relative paths or absolute paths interchangeably

Because it appears Github actions run under a different process/directory than the tests are defined in. This checks to use the location that the process is using for git hub actions, so that it can find proper files.