add RS key validation and grants
add the ability to add a RS key (either public or private) via configuration, file or a directory of files.
each key added has an id that can be selected using a "kid" in a JWT header
instructions on how to add the keys are commented in `common.go`, in short:

this adds the following config fields:
1. **token_rs_key** - (public/private RSA key (PEM encoded with escaped newlines))
2. **token_rs_keys** - a mapping of kid's and key values
3. **token_rs_file** - a path to a PEM encoded file
4. **token_rs_files** - a mapping of kid's to PEM encoded files
5. **token_rs_dir** - a directory of PEM files, where the name/path is the kid

this also adds env vars:
1. **JWT_RS_KEY** - the PEM encoded key
2. **JWT_RS_FILE** - the path to a PEM encoded file
_Both of the above env vars can take a suffix as a kid (i.e. **JWT_RS_KEY_KID_1** where "kid_1" is the kid)
3. **JWT_RS_DIR** - the path to a directory PEM encoded files, and the kid is the path/dir structure

All of the KIDs from env vars are lowercased, where as in the config path, they are case sensitive

note the committed keys are for testing purposes only, and should not be used in production.

Resolves #1